Editor client: Fix ref resource name popup
